Updated 2.80 Script (Meshes Only)
Successfully updated part of the script to be compatible with Blender 2.83.3 LTS and most likely any other Blender version from 2.80.0 and above.

The plugin can now be installed without error and can import a mesh from a PIX CSV file correctly. Incomplete/unedited parts of the script are commented out, as they will be worked on later. Some maintenance and code cleanup was applied, namely setting the script to know what column of data is used for what purpose instead of assuming a specific column will always yield the same type of data.