Minor Tweaks
Just a few minor additions to the code after making a PR to the master branch for the script's original repository. I made some tweaks there, and I decided to update my version as well.

I don't think I need to update the version number for this update, as nothing in terms of functionality was changed. I'll leave that for the bigger stuff ;)