2.7.8 Adjustments + Some Minor Tweaks
Adjusted the old 2.7.8 script to use the custom dynamic column sorter that I implemented in the newer script. Also added in some tweaks based on the PR I made to the original repository as well as format a few things I missed earlier in the newer script. Consistency is key! XD