DirectX PIX Support
This is kind of an optional rollback update, if you wish. There is now a little checkbox to tell the plugin to assume a given CSV file is laid out like a DirectX PIX-generated CSV file. Mainly useful if you're trying to import the output CSV data of a mesh in Renderdoc that has headers like gl_position, v_varying, etc.; or if you just want to look at the warped meshes that you'd normally see when using the original script :P

By default, it assumes all CSV data is from RenderDoc and that the data comes from the Input section when looking at a model's draw calls. If you try to import a non-Input type CSV file, an error will be thrown (may resort to a message box later on).