Added Auto-Cleanup
Added functionality for automatically removing doubles and interior faces.

Before, when a mesh is imported, its tri's were independent of each other, making unnecessary vertices and not making the mesh one solid object. Now, the mesh is a completely solid object when generated instead of many floating triangles.

(I found out that this was an issue when trying to add a smooth modifier to a model in 3DS Max XD)